Refactor project name to enable testing
Changed tool.poetry package name and corresponding directory to ensure working import for tests.
