+++ validateactions/rules/event_trigger.py
@@ -1,8 +1,8 @@
 import yaml
-from lint_problem import LintProblem
+from validateactions.lint_problem import LintProblem
 
 rule = 'event-trigger'
-
+#TODO write tests
 MATCHING_TOKENS = {
     yaml.FlowSequenceStartToken: yaml.FlowSequenceEndToken,
     yaml.BlockSequenceStartToken: yaml.BlockEndToken,
